+++ src/pages/auth/SignIn.tsx
@@ -14,11 +14,7 @@ import PanelHeading from '../../widgets/panel/PanelHeading';
 import Modal from '../../widgets/Modal';
 import PanelBody from '../../widgets/panel/PanelBody';
 import Divider from '../../widgets/Divider';
-
-export interface SignInForm {
-  username: string;
-  password: string;
-}
+import { SignInForm } from '../../types';
 
 const schema = yup
   .object({
