+++ src/store/apis/productApi.ts
@@ -1,5 +1,5 @@
 import { createApi } from '@reduxjs/toolkit/query/react';
-import { Product } from '../models/interfaces';
+import { Product } from '../interfaces/interfaces';
 import baseQueryWithRetry from './baseQueryWithRetry';
 
 const productApi = createApi({
