+++ src/types.ts
@@ -49,3 +49,40 @@ export interface TokenResponse {
   expiresIn: number;
   refreshToken: string;
 }
+
+export interface SignInForm {
+  username: string;
+  password: string;
+}
+
+export interface SignUpForm {
+  username: string;
+  fullName: string;
+  password: string;
+  passwordRepeat: string;
+  code?: string;
+}
+
+export type VerificationStatus = 'PENDING' | 'SENT' | 'COMPLETED' | 'ERROR';
+
+export type VerificationError =
+  | 'INVALID_EMAIL'
+  | 'INVALID_RECAPTCHA'
+  | 'ALREADY_SENT'
+  | 'DUPLICATED'
+  | 'EXPIRED'
+  | 'INVALID_CODE'
+  | null;
+
+export type VerificationState = {
+  status: VerificationStatus;
+  error: VerificationError;
+  timeout: number;
+};
+
+export type VerificationAction =
+  | { type: 'SENT' }
+  | { type: 'COMPLETED' }
+  | { type: 'RESET' }
+  | { type: 'RELOADED'; timeout: number }
+  | { type: 'ERROR'; error: VerificationError };
